+++ repositories/web-application-stack/images/php-web-server-tester/app/src/HttpTests/HomepageTest.php
@@ -1,0 +1,27 @@
+<?php
+declare(strict_types=1);
+
+namespace HttpTests;
+
+class HomepageTest extends \PHPUnit\Framework\TestCase
+{
+    public function getTestsTargetUrl(): string
+    {
+        return $_ENV['WEBSERVERTESTER_TESTS_TARGET_URL'];
+    }
+
+    public function testWelcomeMessage(): void
+    {
+        // Arrange
+        $client = new \GuzzleHttp\Client();
+
+        // Act
+        $response = $client->request('GET', $this->getTestsTargetUrl());
+        $payload = json_decode($response->getBody()->getContents(), true);
+
+        // Assert
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals('application/json', array_shift($response->getHeader('Content-Type')));
+        $this->assertEquals('Hello world!', $payload['message']);
+    }
+}